recode/requirements.txt
jishaku
discord.py[speed]
prodiapy
openai

